rename to `common_types` and move to root namespace
