+++ cot/src/form.rs
@@ -22,7 +22,6 @@
 
 /// Built-in form fields that can be used in a form.
 pub mod fields;
-pub mod types;
 
 use std::borrow::Cow;
 use std::fmt::{Debug, Display};
