+++ cot/src/form/fields.rs
@@ -12,7 +12,7 @@ use cot::db::Auto;
 use crate::auth::PasswordHash;
 #[cfg(feature = "db")]
 use crate::db::LimitedString;
-use crate::form::types::{Email, Password};
+use crate::common_types::{Email, Password};
 use crate::form::{AsFormField, FormField, FormFieldOptions, FormFieldValidationError};
 use crate::html::HtmlTag;
